Added optional creation of a metadata table containing a mapping for SQLite column data types to FDO data types. This is useful if we want to present to the FDO client app the same FDO schema that it used to create the SQLite database. Before, if you ApplySchema with a column that's Int32, and then DescribeSchema, the property would be returned as Int64, since in SQLite all integers are 64 bit.
Note that creation of such table will only be triggered if the connection property "UseFdoMetadata" is set to true when creating the database and when using the database. If such metadata table does not exist, the standard mapping between SQLite and FDO data types that is currently implemented is used.

The definition of the table is as follows:

CREATE TABLE fdo_columns (f_table_name TEXT,f_column_name TEXT,fdo_data_type INTEGER);


git-svn-id: http://svn.osgeo.org/fdo/trunk@4520 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SQLite/Src/Provider/PropertyNameIndex.h
@@ -229,7 +229,7 @@ public:
         {
             sz = children.size();
 
-            for (int i=0; i<sz; i++)
+            for (size_t i=0; i<sz; i++)
                 children[i].InitSize();
 
             if (sz)
